+++ Runtime/Framework/ComposableBehaviour.cs
@@ -12,7 +12,7 @@ namespace DT.UniStart {
     /// If it still doesn't exist, add it to the game object and cache it.
     /// </summary>
     public T GetOrAddComponent<T>() where T : Component {
-      return this.ioc.Get<T>() ?? this.ioc.Add<T>(this.gameObject.GetComponent<T>() ?? this.gameObject.AddComponent<T>());
+      return this.ioc.TryGet<T>() ?? this.ioc.Add<T>(this.gameObject.GetComponent<T>() ?? this.gameObject.AddComponent<T>());
     }
 
     #region Helper Methods for IWatchable